tty: serial: 8250: omap: growable ring-buffer with watchdog
The ring buffer used for cyclic dma transfers will grow according
to the baud rate to allow at least 1 second of buffering. This way
a large (367KiB) would only be created if very high baud rates are
used (3Mbit).

A hrtimer is used as a watchdog to check that the ring buffer is
being processed. If the watchdog triggers and sees there is
unprocessed data in the ring buffer, the DMA engine is paused.

If the watchdog sees that it itself experienced a latency such
that the ring buffer could have become more than half full, a
ring buffer overflow is assumed.

Signed-off-by: John Ogness <john.ogness@linutronix.de>