tty: serial: 8250: omap: convert to using cyclic transfers
By using cyclic transfers it is possible to sustain long running
DMA transfers at 3Mbit, even without flow control.

The ring buffer used for cyclic dma transfers will grow according
to the baud rate to allow 1 second of buffering. This way a large
ring buffer would only be allocated if very high baud rates are
used.

A hrtimer is used as a watchdog to monitor that data in the ring
buffer is not overwritten before being processed by the tty layer.

Using timeout (RDI) interrupts together with DMA has shown to
cause spurious interrupts as well as DMA stalls. To avoid this,
when DMA is used timeout interrupts are not enabled. The hrtimer
watchdog also monitors the RX FIFO in order to simulate the
timeout interrupts. However, since the hrtimer runs at a 1 second
interval, this means data bursts below the RX DMA threshold
(48 bytes) can experience up to 1 second latencies. This makes
using DMA for the console inpractical.

Note that the name of the DMA driver is strcmp'd to detect if
the sDMA engine is being used. This is necessary because the sDMA
driver does not implement pause/resume correctly and therefore
requires slightly different logic for the UART driver.

Signed-off-by: John Ogness <john.ogness@linutronix.de>